Improve Async Paths
